Add FastScanner and IntList
FastScanner can be used as a drop-in replacement for the standard java scanner, and adding new types of tokens to it isn't much harder than regex. Some examples can be seen in the code itself, I will release better docs later.

IntList is a space-efficient implementation for List<Integer>, storing a basic array of primitive ints and resizing it as needed. Complexity is amortized O(1)